+++ src/main/java/org/groupscope/entity/Learner.java
@@ -4,6 +4,7 @@ package org.groupscope.entity;
 import org.hibernate.annotations.Type;
 
 import javax.persistence.*;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
@@ -17,7 +18,7 @@ public class Learner<T> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
@@ -40,13 +41,14 @@ public class Learner<T> {
     private List<Task<TaskType>> tasks;
 
     public Learner() {
+        //tasks = new ArrayList<>();
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
